third_party/blink/web_tests/external/wpt/custom-elements/reactions/HTMLButtonElement-expected.txt

This is a testharness.js-based test.
FAIL autofocus on HTMLButtonElement must enqueue an attributeChanged reaction when adding autofocus content attribute assert_equals: expected "true" but got ""
FAIL autofocus on HTMLButtonElement must enqueue an attributeChanged reaction when replacing an existing attribute assert_equals: expected "true" but got ""
FAIL disabled on HTMLButtonElement must enqueue an attributeChanged reaction when adding disabled content attribute assert_equals: expected "true" but got ""
FAIL disabled on HTMLButtonElement must enqueue an attributeChanged reaction when replacing an existing attribute assert_equals: expected "true" but got ""
PASS name on HTMLButtonElement must enqueue an attributeChanged reaction when adding name content attribute
PASS name on HTMLButtonElement must enqueue an attributeChanged reaction when replacing an existing attribute
PASS value on HTMLButtonElement must enqueue an attributeChanged reaction when adding value content attribute
PASS value on HTMLButtonElement must enqueue an attributeChanged reaction when replacing an existing attribute
PASS type on HTMLButtonElement must enqueue an attributeChanged reaction when adding a new attribute
PASS type on HTMLButtonElement must enqueue an attributeChanged reaction when replacing an existing attribute
PASS formAction on HTMLButtonElement must enqueue an attributeChanged reaction when adding a new attribute
PASS formAction on HTMLButtonElement must enqueue an attributeChanged reaction when replacing an existing attribute
PASS formEnctype on HTMLButtonElement must enqueue an attributeChanged reaction when adding a new attribute
PASS formEnctype on HTMLButtonElement must enqueue an attributeChanged reaction when replacing an existing attribute
PASS formMethod on HTMLButtonElement must enqueue an attributeChanged reaction when adding a new attribute
PASS formMethod on HTMLButtonElement must enqueue an attributeChanged reaction when replacing an existing attribute
FAIL formNoValidate on HTMLButtonElement must enqueue an attributeChanged reaction when adding a new attribute assert_equals: expected "true" but got ""
FAIL formNoValidate on HTMLButtonElement must enqueue an attributeChanged reaction when replacing an existing attribute assert_equals: expected "true" but got ""
PASS formTarget on HTMLButtonElement must enqueue an attributeChanged reaction when adding a new attribute
PASS formTarget on HTMLButtonElement must enqueue an attributeChanged reaction when replacing an existing attribute
Harness: the test ran to completion.
